Add styling to inputs
This visually groups labels and their inputs as well as removes most all
user-agent styling from said inputs. This has been tested with Firefox
and Chrome and should work with all webkit-based browsers.

While this focuses mostly on the styling of these elements, a lot of the
underlying HTML structure was changed to suit a more accessible style
that also accommodates for these changes. It delves slightly into the
admin-specific code to fix up a few things related to input but doesn't
go too far because that code is a mess and needs a whole PR of its own.