Break subtyped packets into individual packets
This splits the current ACK packet and the ChatLookup packet, creating
two new packets in place of each - ACKPlace/ACKUndo and
ChatLookupByMessageId/ChatLookupByUsername.

Regarding the ACK packet, while there is full data structure similarity
between the two, this is incidental and it can be observed that the
previous generalisation of acknowledgement could not be fulfilled by
this packet - see the ACKClientUpdate packet, for example. Rather than
rename the packet to something like `ACKPlaceOrUndo` it seemed more
sensible to instead break them apart since this also decreases the code
complexity when handling packets on the clientside given there is now no
need to analyse a packet further than its type to determine its use.

Regarding the ChatLookup packet, while it may seem like there may be
good reason to implement this as one packet - it fulfills the same or a
substantially similar functional purpose in either mode - splitting it
starts to look like a better idea when realising that the data types and
field names do not match between packets. This manifested previously in
awkward parsing code on the server side with the new code being slightly
cleaner.