Rename most packets to more clearly convey their usage
While adding the verbs to the client packets seems reasonable, the
server packets mostly being prefixed with 'Notify' is perhaps
questionable. 'Notify' seemed like the best descriptor for what thay are
since it is usually unknown to the client when they will recieve one and
they contain data the client should know about but shouldn't respond to
the server for. The prefix 'Return' is used similarly except that it can
be expected to be sent in response to a client's request.