Move packets based on prefix and remove prefixes
Since there are no HTTP packet classes that represent a client request,
this new structure seems lopsided. I propose that this simply reflects
the lopsided-ness of the implementation and that it would be appropriate
to have client HTTP data classes alongside those in future.

Regarding the simplification of class names here and the collisions it
begins to introduce, I believe this is indication of poor class names
even after recent renamings. It would not be unreasonable to expect
another round of renaming following this commit.